requirements.txt
streamlit
streamlit-folium
folium
rasterio
shapely
geopandas
pyproj
numpy
matplotlib
boto3
botocore
requests
sentinelsat
ultralytics
opencv-python
Pillow
reportlab
arabic-reshaper
python-bidi
